+++ plone/app/widgets/utils.py
@@ -2,10 +2,13 @@
 
 from Products.CMFCore.interfaces import ISiteRoot
 from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.interfaces import ITypesSchema
+from plone.registry.interfaces import IRegistry
 from datetime import datetime
 from plone.app.layout.navigation.root import getNavigationRootObject
 from zope.component import providedBy
 from zope.component import queryUtility
+from zope.component import getUtility
 from zope.component.hooks import getSite
 from zope.i18n import translate
 from zope.i18nmessageid import MessageFactory
@@ -136,10 +139,9 @@ def get_relateditems_options(context, value, separator, vocabulary_name,
     options.setdefault('homeText', msgstr)
     options.setdefault('folderTypes', ['Folder'])
 
-    properties = getToolByName(context, 'portal_properties')
-    if properties:
-        options['folderTypes'] = properties.site_properties.getProperty(
-            'typesLinkToFolderContentsInFC', options['folderTypes'])
+    registry = getUtility(IRegistry)
+    settings = registry.forInterface(ITypesSchema, prefix="plone")
+    options['folderTypes'] = settings.types_link_to_folder_contents
 
     return options