+++ plone/app/widgets/tests/test_dx.py
@@ -202,7 +202,10 @@ class DatetimeWidgetTests(unittest.TestCase):
         self.request = TestRequest(environ={'HTTP_ACCEPT_LANGUAGE': 'en'})
         self.field = Datetime(__name__='datetimefield')
         self.widget = DatetimeWidget(self.request)
-        self.widget.pattern_options = {'date': {'firstDay': 0}}
+        self.widget.pattern_options = {
+            'date': {'firstDay': 0},
+            'time': {'interval': 5}
+        }
 
     def test_widget(self):
         current_year = datetime.today().year